[CHORE] 기존 패키지 경로 재설정
