--- src/main/java/com/tiki/server/common/Constants.java
@@ -1,8 +1,0 @@
-package com.tiki.server.common;
-
-public class Constants {
-    public static final String AUTHORIZATION = "Authorization";
-    public static final String BEARER = "Bearer ";
-    public static final String WRONG_INPUT = "잘못된 JSON 형식입니다.";
-    public static final int INIT_NUM = 0;
-}
