+++ src/main/java/com/tiki/server/email/Constants.java
@@ -1,0 +1,6 @@
+package com.tiki.server.email;
+
+public class Constants {
+    public static final String MAIL_FORMAT_EDU = ".edu";
+    public static final String MAIL_FORMAT_AC_KR = ".ac.kr";
+}
