+++ src/main/java/com/tiki/server/email/emailverification/dto/request/CodeVerificationRequest.java
@@ -1,0 +1,11 @@
+package com.tiki.server.email.emailverification.dto.request;
+
+import lombok.NonNull;
+
+public record CodeVerificationRequest(
+        @NonNull
+        String email,
+        @NonNull
+        String code
+) {
+}
