+++ src/main/java/com/tiki/server/email/emailverification/message/SuccessMessage.java
@@ -1,0 +1,13 @@
+package com.tiki.server.email.emailverification.message;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum SuccessMessage {
+
+    SUCCESS_VALIDATION("인증 성공");
+
+    private final String message;
+}
