--- src/main/java/com/tiki/server/emailverification/constants/EmailConstants.java
@@ -1,16 +1,0 @@
-package com.tiki.server.emailverification.constants;
-
-public class EmailConstants {
-
-    public static final String TIKI_EMAIL = "hello.wer.tiki@gmail.com";
-    public static final String MAIL_SUBJECT_SIGN_UP = "[Ti.Ki] 회원가입: 이메일 인증번호 안내";
-    public static final String MAIL_SUBJECT_CHANGING_PASSWORD = "[Ti.Ki] 회원가입: 이메일 인증번호 안내";
-    public static final String IMG_PATH = "images/mail_logo.png";
-    public static final String MAIL_FORMAT_EDU = ".edu";
-    public static final String MAIL_FORMAT_AC_KR = ".ac.kr";
-    public static final String TEMPLATE_NAME = "certification";
-    public static final int CODE_LENGTH = 6;
-    public static final int CODE_NUM_MAX_VALUE_PER_WORD = 10;
-    public static final String CERTIFICATION_PAGE_LOGO_IMAGE_VAR = "image";
-    public static final String CERTIFICATION_PAGE_CODEE_VAR = "code";
-}
