--- src/main/java/com/tiki/server/emailverification/dto/request/CodeVerificationRequest.java
@@ -1,11 +1,0 @@
-package com.tiki.server.emailverification.dto.request;
-
-import lombok.NonNull;
-
-public record CodeVerificationRequest(
-        @NonNull
-        String email,
-        @NonNull
-        String code
-) {
-}
