--- src/main/java/com/tiki/server/emailverification/repository/EmailVerificationRepository.java
@@ -1,13 +1,0 @@
-package com.tiki.server.emailverification.repository;
-
-import com.tiki.server.emailverification.domain.EmailVerification;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface EmailVerificationRepository extends CrudRepository<EmailVerification, Long> {
-
-    Optional<EmailVerification> findById(String email);
-}
